filterx/filterx-object: add str() method
Just like Python, sometimes we want to turn an object into a string to add
it to trace messages (repr) indended to the developer, in other cases we
want a simpler string representation aimed at customers (str).

This patch adds str() in addition to the existing repr().  If str() is not
implemented it defaults to using repr, so defaults doing the same as today.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>